[refactor] パーサのコア部分を移動
